DescriptionsState reads data from a file, as verified through an iostream test
Currently, the file reading code for this purpose (in DataTables.cpp) is the same code (except in separate functions) as the code for reading the rankings file. This may suggest a need to change the way the rankings are obtained (for example, to use the DescriptionData to obtain the RankingData).